Reworked Content a bit and working on the chatConfigSpace starting with contacts.

+++ library/src/main/java/net/sharksystem/sharknet_api_android/dummy_impl/ImplContent.java
@@ -153,7 +153,7 @@ public class ImplContent implements Content {
 	}
 
 	@Override
-	public int getLength() {
+	public long getLength() {
 		return (int) this.sharkFile.getContentLength();
 	}
